pyproject.toml
[tool.poetry]
name = "housing_prices"
version = "0.1.0"
description = "Faroese Housing Prices"
authors = ["Rogvi David Arge <rogvidarge@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.2"
jupyterlab = "^3.1.10"
beautifulsoup4 = "^4.9.3"
html5lib = "^1.1"
lxml = "^4.6.3"
python-dotenv = "^0.19.0"
browser-cookie3 = "^0.12.1"
holoviews = "^1.14.5"
bokeh = "^2.3.3"
panel = "^0.12.1"
geopandas = "^0.9.0"
ipython-sql = "^0.4.0"
SQLAlchemy = "^1.4.23"
datashader = "^0.13.0"
tqdm = "^4.62.3"
ipywidgets = "^7.6.5"
jupyterlab-widgets = "^1.0.2"
matplotlib = "^3.4.3"
hvplot = "^0.8.1"
httpx = "^0.23.3"
asyncio = "^3.4.3"
minio = "^7.1.14"
dvc = "^2.51.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
